+++ haystack/nodes/prompt/invocation_layer/hugging_face.py
@@ -21,9 +21,9 @@ with LazyImport(message="Run 'pip install farm-haystack[inference]'") as torch_a
         GenerationConfig,
         Pipeline,
     )
-    from huggingface_hub import model_info
     from haystack.modeling.utils import initialize_device_settings  # pylint: disable=ungrouped-imports
     from haystack.nodes.prompt.invocation_layer.handlers import HFTokenStreamingHandler
+    from haystack.nodes.prompt.invocation_layer.utils import get_task
 
     class StopWordsCriteria(StoppingCriteria):
         """
@@ -42,15 +42,6 @@ with LazyImport(message="Run 'pip install farm-haystack[inference]'") as torch_a
         def __call__(self, input_ids: torch.LongTensor, scores: torch.FloatTensor, **kwargs) -> bool:
             stop_result = torch.isin(self.stop_words["input_ids"], input_ids[-1])
             return any(all(stop_word) for stop_word in stop_result)
-
-    def get_task(model: str, use_auth_token: Optional[Union[str, bool]] = None, timeout: float = 3.0) -> Optional[str]:
-        """
-        Simplified version of transformers.pipelines.get_task with support for timeouts
-        """
-        try:
-            return model_info(model, token=use_auth_token, timeout=timeout).pipeline_tag
-        except Exception as e:
-            raise RuntimeError(f"The task of {model} could not be checked because of the following error: {e}") from e
 
 
 class HFLocalInvocationLayer(PromptModelInvocationLayer):
